Revert "Revert "[Rollout] Production rollout, 2023-05-10""
This reverts commit 014183c82411e81495b3845976d1bc49aaad1fe7, reversing
changes made to 98e989a1a1160b5d8bb81330c7f551b2c5116f76.
