+++ arcade-services.sln
@@ -125,10 +125,6 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "DotNet.Status.Web.Tests", "
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Microsoft.DotNet.DarcLib.Tests", "src\Microsoft.DotNet.Darc\tests\Microsoft.DotNet.DarcLib.Tests\Microsoft.DotNet.DarcLib.Tests.csproj", "{F78EA40D-7926-4B38-AB4A-ED1E85FAF32C}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "CoreHealthMonitor", "src\Maestro\CoreHealthMonitor\CoreHealthMonitor.csproj", "{46B34C5F-FA4E-4D6C-B94E-C31F6A158302}"
-EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "CoreHealthMonitor.Tests", "src\Maestro\CoreHealthMonitor.Tests\CoreHealthMonitor.Tests.csproj", "{024A6AFE-FC96-4527-AA77-15BC7CDCD809}"
-EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Microsoft.DotNet.Maestro.Tasks.Tests", "src\Maestro\Microsoft.DotNet.Maestro.Tasks.Tests\Microsoft.DotNet.Maestro.Tasks.Tests.csproj", "{C817E57B-319A-42A0-B5B8-C65E4A777861}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Microsoft.DncEng.SecretManager", "src\Microsoft.DncEng.SecretManager\Microsoft.DncEng.SecretManager.csproj", "{56FC5D20-350C-4F52-B862-1FCDCA562DB0}"
@@ -152,6 +148,8 @@ EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Maestro.MergePolicyEvaluation", "src\Maestro\Maestro.MergePolicyEvaluation\Maestro.MergePolicyEvaluation.csproj", "{D16DB73B-4EDF-4E61-9392-238CF3800DE2}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "AzureDevOpsClient.Tests", "src\Telemetry\AzureDevOpsClient.Tests\AzureDevOpsClient.Tests.csproj", "{6CAC1E0D-C81E-4A97-A836-2FEB697C424D}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "MaestroCoreHealthMonitor", "src\Maestro\MaestroCoreHealthMonitor\MaestroCoreHealthMonitor.csproj", "{D0C0C883-C49E-43BD-9E0B-65C6656DB314}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -607,30 +605,6 @@ Global
 		{F78EA40D-7926-4B38-AB4A-ED1E85FAF32C}.Release|x64.Build.0 = Release|Any CPU
 		{F78EA40D-7926-4B38-AB4A-ED1E85FAF32C}.Release|x86.ActiveCfg = Release|Any CPU
 		{F78EA40D-7926-4B38-AB4A-ED1E85FAF32C}.Release|x86.Build.0 = Release|Any CPU
-		{46B34C5F-FA4E-4D6C-B94E-C31F6A158302}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{46B34C5F-FA4E-4D6C-B94E-C31F6A158302}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{46B34C5F-FA4E-4D6C-B94E-C31F6A158302}.Debug|x64.ActiveCfg = Debug|Any CPU
-		{46B34C5F-FA4E-4D6C-B94E-C31F6A158302}.Debug|x64.Build.0 = Debug|Any CPU
-		{46B34C5F-FA4E-4D6C-B94E-C31F6A158302}.Debug|x86.ActiveCfg = Debug|Any CPU
-		{46B34C5F-FA4E-4D6C-B94E-C31F6A158302}.Debug|x86.Build.0 = Debug|Any CPU
-		{46B34C5F-FA4E-4D6C-B94E-C31F6A158302}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{46B34C5F-FA4E-4D6C-B94E-C31F6A158302}.Release|Any CPU.Build.0 = Release|Any CPU
-		{46B34C5F-FA4E-4D6C-B94E-C31F6A158302}.Release|x64.ActiveCfg = Release|Any CPU
-		{46B34C5F-FA4E-4D6C-B94E-C31F6A158302}.Release|x64.Build.0 = Release|Any CPU
-		{46B34C5F-FA4E-4D6C-B94E-C31F6A158302}.Release|x86.ActiveCfg = Release|Any CPU
-		{46B34C5F-FA4E-4D6C-B94E-C31F6A158302}.Release|x86.Build.0 = Release|Any CPU
-		{024A6AFE-FC96-4527-AA77-15BC7CDCD809}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{024A6AFE-FC96-4527-AA77-15BC7CDCD809}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{024A6AFE-FC96-4527-AA77-15BC7CDCD809}.Debug|x64.ActiveCfg = Debug|Any CPU
-		{024A6AFE-FC96-4527-AA77-15BC7CDCD809}.Debug|x64.Build.0 = Debug|Any CPU
-		{024A6AFE-FC96-4527-AA77-15BC7CDCD809}.Debug|x86.ActiveCfg = Debug|Any CPU
-		{024A6AFE-FC96-4527-AA77-15BC7CDCD809}.Debug|x86.Build.0 = Debug|Any CPU
-		{024A6AFE-FC96-4527-AA77-15BC7CDCD809}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{024A6AFE-FC96-4527-AA77-15BC7CDCD809}.Release|Any CPU.Build.0 = Release|Any CPU
-		{024A6AFE-FC96-4527-AA77-15BC7CDCD809}.Release|x64.ActiveCfg = Release|Any CPU
-		{024A6AFE-FC96-4527-AA77-15BC7CDCD809}.Release|x64.Build.0 = Release|Any CPU
-		{024A6AFE-FC96-4527-AA77-15BC7CDCD809}.Release|x86.ActiveCfg = Release|Any CPU
-		{024A6AFE-FC96-4527-AA77-15BC7CDCD809}.Release|x86.Build.0 = Release|Any CPU
 		{C817E57B-319A-42A0-B5B8-C65E4A777861}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{C817E57B-319A-42A0-B5B8-C65E4A777861}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{C817E57B-319A-42A0-B5B8-C65E4A777861}.Debug|x64.ActiveCfg = Debug|Any CPU
@@ -763,6 +737,18 @@ Global
 		{6CAC1E0D-C81E-4A97-A836-2FEB697C424D}.Release|x64.Build.0 = Release|Any CPU
 		{6CAC1E0D-C81E-4A97-A836-2FEB697C424D}.Release|x86.ActiveCfg = Release|Any CPU
 		{6CAC1E0D-C81E-4A97-A836-2FEB697C424D}.Release|x86.Build.0 = Release|Any CPU
+		{D0C0C883-C49E-43BD-9E0B-65C6656DB314}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{D0C0C883-C49E-43BD-9E0B-65C6656DB314}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{D0C0C883-C49E-43BD-9E0B-65C6656DB314}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{D0C0C883-C49E-43BD-9E0B-65C6656DB314}.Debug|x64.Build.0 = Debug|Any CPU
+		{D0C0C883-C49E-43BD-9E0B-65C6656DB314}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{D0C0C883-C49E-43BD-9E0B-65C6656DB314}.Debug|x86.Build.0 = Debug|Any CPU
+		{D0C0C883-C49E-43BD-9E0B-65C6656DB314}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{D0C0C883-C49E-43BD-9E0B-65C6656DB314}.Release|Any CPU.Build.0 = Release|Any CPU
+		{D0C0C883-C49E-43BD-9E0B-65C6656DB314}.Release|x64.ActiveCfg = Release|Any CPU
+		{D0C0C883-C49E-43BD-9E0B-65C6656DB314}.Release|x64.Build.0 = Release|Any CPU
+		{D0C0C883-C49E-43BD-9E0B-65C6656DB314}.Release|x86.ActiveCfg = Release|Any CPU
+		{D0C0C883-C49E-43BD-9E0B-65C6656DB314}.Release|x86.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -801,13 +787,12 @@ Global
 		{B3912147-0659-4960-83EE-BACCC24AFB6D} = {AE791E26-78E1-4936-BCF4-1BF5152CBBD6}
 		{2238029C-6647-4A3D-BF26-69D33BA152A0} = {AE791E26-78E1-4936-BCF4-1BF5152CBBD6}
 		{F78EA40D-7926-4B38-AB4A-ED1E85FAF32C} = {C0B90032-033F-4C3E-809F-BD3339E1E313}
-		{46B34C5F-FA4E-4D6C-B94E-C31F6A158302} = {AE791E26-78E1-4936-BCF4-1BF5152CBBD6}
-		{024A6AFE-FC96-4527-AA77-15BC7CDCD809} = {AE791E26-78E1-4936-BCF4-1BF5152CBBD6}
 		{C817E57B-319A-42A0-B5B8-C65E4A777861} = {AE791E26-78E1-4936-BCF4-1BF5152CBBD6}
 		{A8FF438E-59EF-4A93-9DBF-1970FE1368D7} = {0D9FE592-7EB7-4CEE-95E8-3379B98041C6}
 		{8B2E07FA-ED55-40B4-8D3B-F73B4BD641F8} = {C0B90032-033F-4C3E-809F-BD3339E1E313}
 		{D16DB73B-4EDF-4E61-9392-238CF3800DE2} = {AE791E26-78E1-4936-BCF4-1BF5152CBBD6}
 		{6CAC1E0D-C81E-4A97-A836-2FEB697C424D} = {0D9FE592-7EB7-4CEE-95E8-3379B98041C6}
+		{D0C0C883-C49E-43BD-9E0B-65C6656DB314} = {AE791E26-78E1-4936-BCF4-1BF5152CBBD6}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {32B9C883-432E-4FC8-A1BF-090EB033DD5B}
