+++ eng/Versions.props
@@ -15,27 +15,28 @@
     <MicrosoftDotNetGitIssueManagerVersion>6.0.0-beta.23221.7</MicrosoftDotNetGitIssueManagerVersion>
     <MicrosoftDotNetVersionToolsVersion>6.0.0-beta.23221.7</MicrosoftDotNetVersionToolsVersion>
     <MicrosoftNetTestSdkVersion>17.4.1</MicrosoftNetTestSdkVersion>
-    <MicrosoftDotNetInternalLoggingVersion>1.1.0-beta.23221.1</MicrosoftDotNetInternalLoggingVersion>
-    <MicrosoftAspNetCoreApiPaginationVersion>1.1.0-beta.23221.1</MicrosoftAspNetCoreApiPaginationVersion>
-    <MicrosoftAspNetCoreApiVersioningVersion>1.1.0-beta.23221.1</MicrosoftAspNetCoreApiVersioningVersion>
-    <MicrosoftAspNetCoreApiVersioningAnalyzersVersion>1.1.0-beta.23221.1</MicrosoftAspNetCoreApiVersioningAnalyzersVersion>
-    <MicrosoftAspNetCoreApiVersioningSwashbuckleVersion>1.1.0-beta.23221.1</MicrosoftAspNetCoreApiVersioningSwashbuckleVersion>
-    <MicrosoftDncEngCommandLineLibVersion>1.1.0-beta.23221.1</MicrosoftDncEngCommandLineLibVersion>
-    <MicrosoftDncEngConfigurationExtensionsVersion>1.1.0-beta.23221.1</MicrosoftDncEngConfigurationExtensionsVersion>
-    <MicrosoftDotNetAuthenticationAlgorithmsVersion>1.1.0-beta.23221.1</MicrosoftDotNetAuthenticationAlgorithmsVersion>
-    <MicrosoftDotNetGitHubAuthenticationVersion>1.1.0-beta.23221.1</MicrosoftDotNetGitHubAuthenticationVersion>
-    <MicrosoftDotNetInternalDependencyInjectionTestingVersion>1.1.0-beta.23221.1</MicrosoftDotNetInternalDependencyInjectionTestingVersion>
-    <MicrosoftDotNetInternalDependencyInjectionVersion>1.1.0-beta.23221.1</MicrosoftDotNetInternalDependencyInjectionVersion>
-    <MicrosoftDotNetInternalHealthVersion>1.1.0-beta.23221.1</MicrosoftDotNetInternalHealthVersion>
-    <MicrosoftDotNetInternalTestingDependencyInjectionAbstractionsVersion>1.1.0-beta.23221.1</MicrosoftDotNetInternalTestingDependencyInjectionAbstractionsVersion>
-    <MicrosoftDotNetInternalTestingDependencyInjectionCodeGenVersion>1.1.0-beta.23221.1</MicrosoftDotNetInternalTestingDependencyInjectionCodeGenVersion>
-    <MicrosoftDotNetInternalTestingUtilityVersion>1.1.0-beta.23221.1</MicrosoftDotNetInternalTestingUtilityVersion>
-    <MicrosoftDotNetKustoVersion>1.1.0-beta.23221.1</MicrosoftDotNetKustoVersion>
-    <MicrosoftDotNetMetricsVersion>1.1.0-beta.23221.1</MicrosoftDotNetMetricsVersion>
-    <MicrosoftDotNetServiceFabricServiceHostVersion>1.1.0-beta.23221.1</MicrosoftDotNetServiceFabricServiceHostVersion>
-    <MicrosoftDotNetServicesUtilityVersion>1.1.0-beta.23221.1</MicrosoftDotNetServicesUtilityVersion>
-    <MicrosoftDotNetWebAuthenticationVersion>1.1.0-beta.23221.1</MicrosoftDotNetWebAuthenticationVersion>
-    <ServiceFabricMocksVersion>1.1.0-beta.23221.1</ServiceFabricMocksVersion>
+    <MicrosoftDotNetInternalLoggingVersion>1.1.0-beta.23253.1</MicrosoftDotNetInternalLoggingVersion>
+    <MicrosoftAspNetCoreApiPaginationVersion>1.1.0-beta.23253.1</MicrosoftAspNetCoreApiPaginationVersion>
+    <MicrosoftAspNetCoreApiVersioningVersion>1.1.0-beta.23253.1</MicrosoftAspNetCoreApiVersioningVersion>
+    <MicrosoftAspNetCoreApiVersioningAnalyzersVersion>1.1.0-beta.23253.1</MicrosoftAspNetCoreApiVersioningAnalyzersVersion>
+    <MicrosoftAspNetCoreApiVersioningSwashbuckleVersion>1.1.0-beta.23253.1</MicrosoftAspNetCoreApiVersioningSwashbuckleVersion>
+    <MicrosoftDncEngCommandLineLibVersion>1.1.0-beta.23253.1</MicrosoftDncEngCommandLineLibVersion>
+    <MicrosoftDncEngConfigurationExtensionsVersion>1.1.0-beta.23253.1</MicrosoftDncEngConfigurationExtensionsVersion>
+    <MicrosoftDotNetAuthenticationAlgorithmsVersion>1.1.0-beta.23253.1</MicrosoftDotNetAuthenticationAlgorithmsVersion>
+    <MicrosoftDotNetGitHubAuthenticationVersion>1.1.0-beta.23253.1</MicrosoftDotNetGitHubAuthenticationVersion>
+    <MicrosoftDotNetInternalDependencyInjectionTestingVersion>1.1.0-beta.23253.1</MicrosoftDotNetInternalDependencyInjectionTestingVersion>
+    <MicrosoftDotNetInternalDependencyInjectionVersion>1.1.0-beta.23253.1</MicrosoftDotNetInternalDependencyInjectionVersion>
+    <MicrosoftDotNetInternalHealthVersion>1.1.0-beta.23253.1</MicrosoftDotNetInternalHealthVersion>
+    <MicrosoftDotNetInternalTestingDependencyInjectionAbstractionsVersion>1.1.0-beta.23253.1</MicrosoftDotNetInternalTestingDependencyInjectionAbstractionsVersion>
+    <MicrosoftDotNetInternalTestingDependencyInjectionCodeGenVersion>1.1.0-beta.23253.1</MicrosoftDotNetInternalTestingDependencyInjectionCodeGenVersion>
+    <MicrosoftDotNetInternalTestingUtilityVersion>1.1.0-beta.23253.1</MicrosoftDotNetInternalTestingUtilityVersion>
+    <MicrosoftDotNetKustoVersion>1.1.0-beta.23253.1</MicrosoftDotNetKustoVersion>
+    <MicrosoftDotNetMetricsVersion>1.1.0-beta.23253.1</MicrosoftDotNetMetricsVersion>
+    <MicrosoftDotNetServiceFabricServiceHostVersion>1.1.0-beta.23253.1</MicrosoftDotNetServiceFabricServiceHostVersion>
+    <MicrosoftDotNetServicesUtilityVersion>1.1.0-beta.23253.1</MicrosoftDotNetServicesUtilityVersion>
+    <MicrosoftDotNetWebAuthenticationVersion>1.1.0-beta.23253.1</MicrosoftDotNetWebAuthenticationVersion>
+    <ServiceFabricMocksVersion>1.1.0-beta.23253.1</ServiceFabricMocksVersion>
+    <CoreHealthMonitorVersion>1.1.0-beta.23253.1</CoreHealthMonitorVersion>
   </PropertyGroup>
   <!--Package names-->
   <PropertyGroup>
