+++ src/Maestro/MaestroCoreHealthMonitor/PackageRoot/ServiceManifest.xml
@@ -14,7 +14,7 @@
   <CodePackage Name="Code" Version="1.0.42">
     <EntryPoint>
       <ExeHost>
-        <Program>CoreHealthMonitor.exe</Program>
+        <Program>MaestroCoreHealthMonitor.exe</Program>
         <ConsoleRedirection FileRetentionCount="5" FileMaxSizeInKb="2048"/>
       </ExeHost>
     </EntryPoint>
@@ -23,7 +23,7 @@
     </EnvironmentVariables>
   </CodePackage>
 
-  <!-- Config package is the contents of the Config directoy under PackageRoot that contains an 
+  <!-- Config package is the contents of the Config directory under PackageRoot that contains an 
        independently-updateable and versioned set of custom configuration settings for your service. -->
   <ConfigPackage Name="Config" Version="1.0.42" />
